Preventing SignalRFuture's cancel handlers from running multiple times
Might be useful because the connection future's cancel() method can both
be called internally in the library if an error happens during the
connection process, as well as from the client program if it decides to
give up with the connection process for some reasons, and both cases
could happen simultaneously, while the cancel handlers (for instance the
one defined in JavaHttpConnection) are not supposed to be called
multiple times. Letting them running multiple times wouldn't be a
serious issue though.