Added method to set a handler for the ConnectionTimeOut event
In addition to the ConnectionSlow and ConnectionClosed events, a
ConnectionTimeOut event is now generated when the HeartbeatMonitor
detects a timeout.

That event is generated just before the connection is closed or
restarted as a result of the timeout (and therefore before the
ConnectionClosed event is generated), and it is specified in the
ConnectionBase interface that these operations must be performed by the
implementing classes, so that the handler for the ConnectionTimeOut
event doesn't need to close or restart the connection as it will be
performed automatically when the handler is done running.